add state machine handler from pairing to long press
