+++ adapter/ble/ble.c
@@ -40,16 +40,16 @@ void ble_state_set(app_state cur_state)
     }
 
     if((cur_state == app_fast_advertising) || (cur_state == app_slow_advertising)) {
-        print((u8*)&"ble adv", 7);
+        //print((u8*)&"ble adv", 7);
         ble_switch_cb(BLE_ADVERTISE, NULL);
     } else if(cur_state == app_idle){
-        print((u8*)&"ble idle", 8);
+        //print((u8*)&"ble idle", 8);
         ble_switch_cb(BLE_STOP_ADVERTISE, NULL);
     } else if(cur_state == app_connected){
-        print((u8*)&"ble con", 7);
+        //print((u8*)&"ble con", 7);
         ble_switch_cb(BLE_CONNECT, NULL);
     } else {
-        print((u8*)&"ble discon", 10);
+        //print((u8*)&"ble discon", 10);
         ble_switch_cb(BLE_DISCONNECT, NULL);
     }
     ble_last_state = cur_state;
