+++ business/business.c
@@ -19,38 +19,40 @@ static business_t business = {
 };
 
 static state_t state[] = {
-	STATE_FILL(CLOCK,       CLOCK_1_MINUTE,     	CLOCK,       	state_clock),
+	STATE_FILL(CLOCK,               CLOCK_1_MINUTE,     	CLOCK,       	        state_clock),
 	/*battery mode*/
-	STATE_FILL(CLOCK,       BATTERY_LOW,        	LOW_VOLTAGE, 	state_low_voltage),
-	STATE_FILL(LOW_VOLTAGE, BATTERY_NORMAL,     	CLOCK,       	state_clock),
+	STATE_FILL(CLOCK,               BATTERY_LOW,        	LOW_VOLTAGE, 	        state_low_voltage),
+	STATE_FILL(LOW_VOLTAGE,         BATTERY_NORMAL,     	CLOCK,       	        state_clock),
 	/*zero adjust mode*/
-	STATE_FILL(CLOCK,       KEY_A_B_LONG_PRESS, 	ZERO_ADJUST, 	state_zero_adjust),
-	STATE_FILL(ZERO_ADJUST, KEY_A_SHORT_PRESS,  	ZERO_ADJUST, 	state_zero_adjust),
-	STATE_FILL(ZERO_ADJUST, KEY_B_SHORT_PRESS,  	ZERO_ADJUST, 	state_zero_adjust),
-	STATE_FILL(ZERO_ADJUST, KEY_M_SHORT_PRESS,		ZERO_ADJUST, 	state_zero_adjust),
-	STATE_FILL(ZERO_ADJUST, KEY_A_B_LONG_PRESS, 	CLOCK,       	state_clock),
+	STATE_FILL(CLOCK,               KEY_A_B_LONG_PRESS, 	ZERO_ADJUST, 	        state_zero_adjust),
+	STATE_FILL(ZERO_ADJUST,         KEY_A_SHORT_PRESS,  	ZERO_ADJUST, 	        state_zero_adjust),
+	STATE_FILL(ZERO_ADJUST,         KEY_B_SHORT_PRESS,  	ZERO_ADJUST, 	        state_zero_adjust),
+	STATE_FILL(ZERO_ADJUST,         KEY_M_SHORT_PRESS,		ZERO_ADJUST, 	        state_zero_adjust),
+	STATE_FILL(ZERO_ADJUST,         KEY_A_B_LONG_PRESS, 	CLOCK,       	        state_clock),
 	/*ble switch open*/	
-	//STATE_FILL(CLOCK,       KEY_M_LONG_PRESS,   	BLE_SWITCH,  	state_ble_switch),
-	STATE_FILL(CLOCK,       KEY_M_LONG_PRESS,   	CLOCK,  	    state_ble_switch),
+	//STATE_FILL(CLOCK,             KEY_M_LONG_PRESS,   	BLE_SWITCH,  	        state_ble_switch),
+	//STATE_FILL(CLOCK,               KEY_M_LONG_PRESS,   	CLOCK,  	            state_ble_switch),
+	STATE_FILL(CLOCK,               KEY_A_LONG_PRESS,   	CLOCK,  	            state_ble_switch),
 	/*notify*/
-	STATE_FILL(CLOCK,       ANCS_NOTIFY_INCOMING,   NOTIFY_COMING,  state_notify),
+	STATE_FILL(CLOCK,               ANCS_NOTIFY_INCOMING,   NOTIFY_COMING,          state_notify),
 	/*battery & week switch*/	
-	STATE_FILL(CLOCK,       KEY_M_SHORT_PRESS,   	BATTERY_WEEK_SWITCH,  state_battery_week_switch),
+	STATE_FILL(CLOCK,               KEY_M_SHORT_PRESS,   	BATTERY_WEEK_SWITCH,    state_battery_week_switch),
 	/*time adjust*/
-	STATE_FILL(CLOCK,       KEY_B_M_LONG_PRESS,   	TIME_ADJUST,  	state_time_adjust),
+	STATE_FILL(CLOCK,               KEY_B_M_LONG_PRESS,   	TIME_ADJUST,  	        state_time_adjust),
 	/*run test*/
-	STATE_FILL(CLOCK,       KEY_A_B_M_LONG_PRESS,   RUN_TEST,  		state_run_test),
+	STATE_FILL(CLOCK,               KEY_A_B_M_LONG_PRESS,   RUN_TEST,  		        state_run_test),
 	/*ble state change*/
-	STATE_FILL(CLOCK,               BLE_ADVERTISE,      CLOCK,                  state_ble_advertise),
-	STATE_FILL(CLOCK,               BLE_STOP_ADVERTISE, CLOCK,                  state_ble_stop_advertise),
-	STATE_FILL(CLOCK,               BLE_CONNECT,        CLOCK,                  state_ble_connect),
-	STATE_FILL(CLOCK,               BLE_DISCONNECT,     CLOCK,                  state_ble_disconnect),
-	STATE_FILL(CLOCK,               PAIRING_PROC,       PAIRING_INITIATE,       state_ble_pairing),
-	STATE_FILL(PAIRING_INITIATE,    BLE_DISCONNECT,     CLOCK,                  state_ble_disconnect),
-	STATE_FILL(PAIRING_INITIATE,    BLE_ADVERTISE,      CLOCK,                  state_ble_advertise),
-	/*device pairing*/
-	STATE_FILL(PAIRING_INITIATE,    PAIRING_PROC,       PAIRING_INITIATE,       state_ble_pairing),
-	STATE_FILL(PAIRING_INITIATE,    PAIRING_PROC,       CLOCK,                  state_ble_pairing),
+	STATE_FILL(CLOCK,               BLE_ADVERTISE,          CLOCK,                  state_ble_advertise),
+	STATE_FILL(CLOCK,               BLE_STOP_ADVERTISE,     CLOCK,                  state_ble_stop_advertise),
+	STATE_FILL(CLOCK,               BLE_CONNECT,            CLOCK,                  state_ble_connect),
+	STATE_FILL(CLOCK,               BLE_DISCONNECT,         CLOCK,                  state_ble_disconnect),
+	STATE_FILL(CLOCK,               PAIRING_PROC,           PAIRING_INITIATE,       state_ble_pairing),
+	STATE_FILL(PAIRING_INITIATE,    BLE_DISCONNECT,         CLOCK,                  state_ble_disconnect),
+	STATE_FILL(PAIRING_INITIATE,    BLE_ADVERTISE,          CLOCK,                  state_ble_advertise),
+	STATE_FILL(PAIRING_INITIATE,    KEY_A_LONG_PRESS,       CLOCK,                  state_ble_switch),
+	/*ble pairing*/
+	STATE_FILL(PAIRING_INITIATE,    PAIRING_PROC,           PAIRING_INITIATE,       state_ble_pairing),
+	STATE_FILL(PAIRING_INITIATE,    PAIRING_PROC,           CLOCK,                  state_ble_pairing),
 };
 
 static s16 adapter_cb_handler(REPORT_E cb, void *args)
